Fixes #1 t/001-extract-from-pdf.t fails (if gs is not installed?)

+++ lib/Health/SHC/Extract.pm
@@ -1,6 +1,7 @@
 use strict;
 use warnings;
 
+use File::Which;
 # ABSTRACT: Extract the Smart Health Card information from files
 
 use PostScript::Convert;
@@ -103,7 +104,14 @@ sub extract_qr_from_pdf {
     use File::Temp qw/ tempfile tempdir /;
     my ($fh, $output_filename) = tempfile('shctempfileXXXXX', SUFFIX => '.png');
 
-    PostScript::Convert::psconvert($filename, filename => $output_filename, format => 'pnggray');
+    my $ret;
+    eval {
+        my $gs_exec = File::Which::which('gs');
+        $ret = defined $gs_exec ? 1 : 0;
+        PostScript::Convert::psconvert($filename, filename => $output_filename, format => 'pnggray') if $ret;
+    };
+    return () if (!$ret);
+
     my @qrcodes = $self->extract_qr_from_png($output_filename);
 
     unlink $output_filename;
